+++ optimus/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.12.11"
+__version__ = "0.12.12"
 
 
 def safe_commit() -> None:
